requirements.txt
appdirs==1.4.4
asgiref==3.5.2
black==22.8.0
chipwhisperer==5.6.1
click==8.1.3
configobj==5.0.6
Cython==0.29.32
Django==4.1.1
ECPy==1.2.5
fastdtw==0.3.4
image==1.5.33
libusb1==3.0.0
Mako==1.2.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pygubu==0.25
pygubu-designer==0.30
pyserial==3.5
quantiphy==2.18.0
screeninfo==0.8.1
six==1.16.0
sqlparse==0.4.2
tkinter-tooltip==2.1.0
tomli==2.0.1
tqdm==4.64.1
